requirements.txt
discord.py
loguru
pip
GitPython
